requirements.txt
python==3.7.5
Cython==0.29.30
cython-bbox==0.1.3
decord==0.6.0
mindspore-gpu==1.6.1
mindspore-serving==1.6.0
ml-collections==0.1.1
matplotlib==3.4.1
motmetrics==1.2.5
numpy==1.21.5
Pillow==9.0.1
PyYAML==6.0
scikit-learn==1.0.2
scipy==1.7.3
pycocotools==2.0